Change amount datatype Double -> Integer

+++ src/main/java/org/reddot15/be_stockmanager/dto/request/ProductUpdateRequest.java
@@ -28,7 +28,7 @@ public class ProductUpdateRequest {
 	Double salePrice;
 
 	@NotNull(message = "Amount is required.")
-	Double amount;
+	Integer amount;
 
 	@NotBlank(message = "Earliest expiry is required.")
 	String earliestExpiry;
