+++ src/main/java/org/reddot15/be_stockmanager/dto/response/ProductRevenueStatResponse.java
@@ -18,7 +18,7 @@ public class ProductRevenueStatResponse {
 
 	String categoryName;
 
-	Double amount;
+	Integer amount;
 
 	Double totalRevenue;
 }
